Refactor file structure, extract validation logic
* to prepare for decoupling validation & db writes, change file structure:
  * move `db/models/donation.js` to `models/donation/schema.js`
  * move `db/dao/donations.js` to `models/donation/dao.js`
  * extract validation methods from dao to `models/donation/validate.js`

* reason for decouple:
  * desired order of operations is (1) validate, (2) submit charge to
   stripe, (3) write to db
  * currently validation happens in db write step
  * to validate separately, must extract validation logic, hence refactor
